requirements.txt
django==2.2.16
djangorestframework==3.11.1
ipython==7.9.0
django-dynamic-fixture==3.1.0
django-test-plus==1.4.0
